+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/zetime/ZeTimeCoordinator.java
@@ -124,7 +124,7 @@ public class ZeTimeCoordinator extends AbstractDeviceCoordinator {
 
     @Override
     public boolean supportsCalendarEvents() {
-        return true;
+        return false;
     }
 
     @Override
@@ -166,6 +166,7 @@ public class ZeTimeCoordinator extends AbstractDeviceCoordinator {
                 R.xml.devicesettings_zetime,
                 R.xml.devicesettings_timeformat,
                 R.xml.devicesettings_wearlocation,
+                R.xml.devicesettings_sync_calendar,
         };
     }
 }